# packages/backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/transit/distance": {
            "get": {
                "description": "Returns the shortest number of stations between an inspector's station and a given user's latitude and longitude coordinates.\nThe distance calculation employs Dijkstra's algorithm to determine the minimal stops required to reach the nearest station from the given coordinates.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "transit"
                ],
                "summary": "Calculate shortest distance to a station",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The station Id of the inspector's current location.",
                        "name": "inspectorStationId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The station Id of the user's current location.",
                        "name": "userStationId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID of the network (defaults to berlin)",
                        "name": "network",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The shortest distance in terms of the number of station stops between the inspector's station and the user's location.",
                        "schema": {
                            "type": "int"
                        }
                    },
                    "400": {
                        "description": "Bad Request: One of the station id parameters is missing.",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found: The specified network does not exist.",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity: One of the stations does not belong to the network.",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "An error occurred in processing the request."
                    }
                }
            }
        },
...
    },
...
}